requirements.txt
absl-py==0.12.0
appdirs==1.4.4
astunparse==1.6.3
attrs==21.2.0
cachetools==4.2.4
cattrs==1.8.0
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
cycler==0.11.0
Cython==0.29.24
dill==0.3.4
dm-tree==0.1.6
ephem==4.1.1
flatbuffers==2.0
fonttools==4.28.2
future==0.18.2
gast==0.4.0
gin-config==0.5.0
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.42.0
h5py==3.6.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.8.2
joblib==1.1.0
kaggle==1.5.12
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
lxml==4.6.4
Markdown==3.3.6
matplotlib==3.5.0
numpy==1.21.4
oauth2client==4.1.3
oauthlib==3.1.1
opencv-python-headless==4.5.4.60
opt-einsum==3.3.0
orbit==0.2
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
portalocker==2.3.2
promise==2.3
protobuf==3.19.1
psutil==5.8.0
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pyephem==9.99
pyparsing==3.0.6
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
requests-cache==0.8.1
requests-oauthlib==1.3.0
rsa==4.7.2
sacrebleu==2.0.0
scikit-learn==1.0.1
scipy==1.7.2
sentencepiece==0.1.96
seqeval==1.2.2
setuptools-scm==6.3.2
six==1.16.0
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-addons==0.15.0
tensorflow-cpu==2.7.0
tensorflow-datasets==4.4.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.22.0
tensorflow-metadata==1.4.0
tensorflow-model-optimization==0.7.0
tensorflow-text==2.7.3
termcolor==1.1.0
text-unidecode==1.3
tf-models-official==2.7.0
tf-slim==1.1.0
threadpoolctl==3.0.0
tomli==1.2.2
tqdm==4.62.3
typeguard==2.13.2
typing_extensions==4.0.0
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
